NILE-292:Refactor side bar menu component
