+++ src/app/theme/components/baSideBar/baSideBar.scss
@@ -156,54 +156,6 @@ $angle-right: "\f101";
     left: $sidebar-width - $sidebar-scroll-width;
   }
 
-  @mixin layout-collapsed() {
-    .al-main {
-      margin-left: 50px;
-    }
-
-    .al-footer {
-      padding-left: 83px
-    }
-  }
-
-  @mixin default-sublist() {
-    padding: 0;
-    list-style: none;
-    position: relative;
-    display: none;
-    &.expanded {
-      display: block;
-    }
-    > ba-menu-item > li {
-      display: block;
-      float: none;
-      padding: 0;
-      border-bottom: none;
-      position: relative;
-      a {
-        display: block;
-        text-shadow: none;
-        font-size: 13px;
-        text-decoration: none;
-        color: $sidebar-text;
-        padding-left: 52px;
-        height: auto;
-        line-height: 29px;
-        &:hover {
-          color: $primary;
-        }
-      }
-      &.selected:not(.with-sub-menu) > a {
-        border: none;
-        background-color: $primary;
-        &:hover {
-          color: $sidebar-text;
-        }
-      }
-
-    }
-  }
-
   .al-sidebar-sublist {
     @include default-sublist();
   }
@@ -225,56 +177,6 @@ $angle-right: "\f101";
     top: 94px;
   }
 
-  @mixin sidebar-collapsed() {
-    .al-sidebar {
-      width: 52px;
-
-      .fa-angle-down, .fa-angle-up {
-        opacity: 0;
-      }
-
-      .al-sidebar-sublist {
-        position: absolute;
-        top: -1px;
-        left: 52px;
-        @include bg-translucent-dark(0.8);
-        width: 0;
-        display: block;
-        overflow: hidden;
-        transition: width 0.5s ease;
-        &.slide-right {
-          width: 135px;
-        }
-        &:before {
-          display: none;
-        }
-        li {
-          &:before {
-            display: none;
-          }
-          a {
-            padding-left: 18px;
-            padding-right: 18px;
-            min-width: 130px;
-            white-space: nowrap;
-          }
-        }
-      }
-
-      .sidebar-hover-elem, .sidebar-select-elem {
-        left: 48px;
-      }
-    }
-  }
-
-  @mixin sidebar-hidden() {
-    .al-sidebar {
-      width: 0;
-    }
-    .sidebar-hover-elem, .sidebar-select-elem {
-      display: none;
-    }
-  }
 
   @mixin sidebar-overlap() {
     .al-sidebar {
@@ -301,44 +203,6 @@ $angle-right: "\f101";
         left: $sidebar-width - 4;
         transition: left 0.5s ease;
       }
-    }
-  }
-
-  .menu-collapsed {
-    .slimScrollBar, .slimScrollRail {
-      display: none !important;
-    }
-  }
-
-  @media (min-width: 1200px) {
-    .menu-collapsed {
-      @include layout-collapsed();
-    }
-  }
-
-  @media (min-width: $resXS + 1) {
-    .menu-collapsed {
-      @include sidebar-collapsed();
-    }
-  }
-
-  @media (max-width: 1200px) and (min-width: $resXS) {
-    @include layout-collapsed();
-  }
-
-  @media (max-width: 1200px) {
-    @include sidebar-overlap();
-  }
-
-  @media (max-width: $resXS) {
-    .menu-collapsed {
-      @include sidebar-hidden();
-    }
-    .al-main {
-      margin-left: 0;
-    }
-    .al-footer {
-      padding-left: 0;
     }
   }
 }
