+++ src/app/theme/components/baSideBar/index.ts
@@ -1,4 +1,4 @@
 /**
  * Created by adediosf on 21/02/17.
  */
-export * from './side-bar.component';
+export * from './baSideBar.component';
