+++ src/app/theme/components/index.ts
@@ -4,9 +4,9 @@
 export * from './baAmChart';
 export * from './baBackTop';
 export * from './baPageTop';
-export * from './side-bar/menu/menu-item';
-export * from './side-bar/menu';
-export * from './side-bar';
+export * from './menu/menu-item';
+export * from './menu';
+export * from './baSideBar';
 export * from './baContentTop';
 export * from './baCard';
 export * from './baChartistChart';
