+++ src/app/theme/components/menu/menu-item/menu-item.component.scss
