+++ src/app/theme/components/menu/menu.component.scss
@@ -1,0 +1,1 @@
+@import '../../sass/conf/conf';
