--- src/app/theme/components/side-bar/menu/menu-item/menu-item.component.scss
@@ -1,1 +1,0 @@
-@import '../../../../sass/conf/conf';
